Move request stats middleware to middlewares
