+++ src/service/http/index.ts
@@ -2,9 +2,9 @@ import { ClientsImplementation, IOClients } from '../../clients/IOClients'
 import { InstanceOptions } from '../../HttpClient'
 import { RouteHandler } from '../typings'
 import { compose } from '../utils/compose'
-import { trackIncomingRequestStats } from '../utils/incomingRequestStats'
 import { clients } from './middlewares/clients'
 import { error } from './middlewares/error'
+import { trackIncomingRequestStats } from './middlewares/requestStats'
 import { removeSetCookie } from './middlewares/setCookie'
 import { timings } from './middlewares/timings'
